Moved pages into directory for DOW
